Basic conversion to typescript for all files. TODO: fix all warnings
