+++ app/.eslintignore
@@ -1,3 +1,3 @@
 # generated code
-src/serviceWorker.js
+src/serviceWorker.ts
 build
