--- app/src/components/App.js
@@ -1,26 +1,0 @@
-import React from 'react';
-import { DrizzleContext } from 'drizzle-react';
-import ArtPieceList from './ArtPieceList';
-import Register from './Register';
-
-const AppComponent = (
-  <DrizzleContext.Consumer>
-    {drizzleContext => {
-      const { drizzle, drizzleState, initialized } = drizzleContext;
-
-      if (!initialized) {
-        // TODO: Show a more user-friendly loading screen
-        return 'Loading...';
-      }
-
-      return (
-        <div>
-          <ArtPieceList />
-          <Register drizzle={drizzle} drizzleState={drizzleState} />
-        </div>
-      );
-    }}
-  </DrizzleContext.Consumer>
-);
-
-export default () => AppComponent;
