+++ app/src/declarations.d.ts
@@ -1,0 +1,5 @@
+declare module 'drizzle'
+declare module 'prop-types'
+declare module 'drizzle-react'
+declare module 'drizzle-react-components'
+declare module '*.json'
