+++ app/tsconfig.json
@@ -1,10 +1,6 @@
 {
   "compilerOptions": {
-    "target": "es6",
-    "outDir": "./dist/",
-    "sourceMap": true,
-    "noImplicitAny": true,
-    "module": "esnext",
+    "target": "es5",
     "lib": [
       "dom",
       "dom.iterable",
@@ -16,6 +12,7 @@
     "allowSyntheticDefaultImports": true,
     "strict": true,
     "forceConsistentCasingInFileNames": true,
+    "module": "esnext",
     "moduleResolution": "node",
     "resolveJsonModule": true,
     "isolatedModules": true,
